refactor(cxx): move LoadFileOrDie() into a new file_utils library
