+++ kythe/cxx/extractor/cxx_extractor.cc
@@ -38,6 +38,7 @@
 #include "clang/Tooling/Tooling.h"
 #include "gflags/gflags.h"
 #include "glog/logging.h"
+#include "kythe/cxx/common/file_utils.h"
 #include "kythe/cxx/common/json_proto.h"
 #include "kythe/cxx/common/kzip_writer.h"
 #include "kythe/cxx/common/path_utils.h"
@@ -1187,21 +1188,6 @@ void MapCompilerResources(clang::tooling::ToolInvocation* invocation,
     llvm::sys::path::append(out_path, file->name);
     invocation->mapVirtualFile(out_path, file->data);
   }
-}
-
-/// \brief Loads all data from a file or terminates the process.
-static std::string LoadFileOrDie(const std::string& file) {
-  FILE* handle = fopen(file.c_str(), "rb");
-  CHECK(handle != nullptr) << "Couldn't open input file " << file;
-  CHECK_EQ(fseek(handle, 0, SEEK_END), 0) << "Couldn't seek " << file;
-  long size = ftell(handle);
-  CHECK_GE(size, 0) << "Bad size for " << file;
-  CHECK_EQ(fseek(handle, 0, SEEK_SET), 0) << "Couldn't seek " << file;
-  std::string content;
-  content.resize(size);
-  CHECK_EQ(fread(&content[0], size, 1, handle), 1) << "Couldn't read " << file;
-  CHECK_NE(fclose(handle), EOF) << "Couldn't close " << file;
-  return content;
 }
 
 void ExtractorConfiguration::SetVNameConfig(const std::string& path) {
