+++ kythe/cxx/extractor/textproto/BUILD
@@ -4,6 +4,7 @@ cc_binary(
     visibility = ["//visibility:public"],
     deps = [
         ":textproto_schema",
+        "//kythe/cxx/common:file_utils",
         "//kythe/cxx/common:kzip_writer",
         "//kythe/cxx/common:path_utils",
         "//kythe/cxx/extractor/proto:lib",
