+++ kythe/cxx/extractor/textproto/textproto_extractor.cc
@@ -30,6 +30,7 @@
 #include "absl/strings/str_split.h"
 #include "gflags/gflags.h"
 #include "glog/logging.h"
+#include "kythe/cxx/common/file_utils.h"
 #include "kythe/cxx/common/kzip_writer.h"
 #include "kythe/cxx/common/path_utils.h"
 #include "kythe/cxx/extractor/proto/proto_extractor.h"
@@ -52,21 +53,6 @@ IndexWriter OpenKzipWriterOrDie(absl::string_view path) {
   auto writer = KzipWriter::Create(path);
   CHECK(writer.ok()) << "Failed to open KzipWriter: " << writer.status();
   return std::move(*writer);
-}
-
-// Loads all data from a file or terminates the process.
-std::string LoadFileOrDie(const std::string& file) {
-  FILE* handle = fopen(file.c_str(), "rb");
-  CHECK(handle != nullptr) << "Couldn't open input file " << file;
-  CHECK_EQ(fseek(handle, 0, SEEK_END), 0) << "Couldn't seek " << file;
-  long size = ftell(handle);
-  CHECK_GE(size, 0) << "Bad size for " << file;
-  CHECK_EQ(fseek(handle, 0, SEEK_SET), 0) << "Couldn't seek " << file;
-  std::string content;
-  content.resize(size);
-  CHECK_EQ(fread(&content[0], size, 1, handle), 1) << "Couldn't read " << file;
-  CHECK_NE(fclose(handle), EOF) << "Couldn't close " << file;
-  return content;
 }
 
 }  // namespace
@@ -107,7 +93,7 @@ Examples:
   CHECK(textproto_args.size() == 1)
       << "Expected 1 textproto file, got " << textproto_args.size();
   std::string textproto_filename = textproto_args[0];
-  const std::string textproto = LoadFileOrDie(textproto_filename);
+  const std::string textproto = ::kythe::LoadFileOrDie(textproto_filename);
 
   const char* output_file = getenv("KYTHE_OUTPUT_FILE");
   CHECK(output_file != nullptr)
